requirements.txt
aiohttp==3.6.1
amqp==2.5.1
async-timeout==3.0.1
attrs==19.1.0
billiard==3.6.1.0
celery==4.3.0
certifi==2019.9.11
chardet==3.0.4
Django==2.2.5
django-environ==0.4.5
djangorestframework==3.10.3
idna==2.8
idna-ssl==1.1.0
importlib-metadata==0.23
kombu==4.6.4
more-itertools==7.2.0
multidict==4.5.2
mysqlclient==1.4.4
pytz==2019.2
redis==3.3.8
requests==2.22.0
slackclient==2.1.0
sqlparse==0.3.0
typing-extensions==3.7.4
urllib3==1.25.4
vine==1.3.0
yarl==1.3.0
zipp==0.6.0
